+++ spikey/__init__.py
@@ -4,7 +4,7 @@ Spikey __init__.
 try:
     # import spikey.<analysis, logger, viz>
     import spikey.core  # import spikey.core.Object
-    from spikey.games import *  # import spikey.<game_type>.<game_class>
+    import spikey.games  # import spikey.games.<game_class>
     from spikey.snn import *  # import spikey.<part_type>.<part_class>
 
     from spikey._metapath_dir_skip import install_skipfolder
