+++ spikey/games/README.md
@@ -5,4 +5,4 @@ can interact with said environment.
 In this simulator they serve as an effective, modular way to give input
 to and interpret feedback from the network.
 
-All games are structured similarly to environments in OpenAI's gym. RL games are nearly exactly the same and should be used to train spiking networks, ideally in a TrainingLoop. MetaRL games are designed for the genetic algorithm(Population) which notably has a game to optimize hyperparameters of a Network-RL experiment.
+All games are structured similarly to environments in OpenAI's gym. RL games are nearly exactly the same and should be used to train spiking networks, ideally in a TrainingLoop.
