+++ unit_tests/test_game.py
@@ -6,7 +6,7 @@ from unit_tests import ModuleTest
 from copy import deepcopy
 from gym.envs.classic_control import cartpole, mountain_car
 from spikey import Key
-from spikey.games import game, RL, MetaRL, gym_wrapper
+from spikey import games
 
 
 class FakeTrainingLoop:
@@ -32,7 +32,7 @@ class TestGame(unittest.TestCase, ModuleTest):
     Tests for games.Game.
     """
 
-    TYPES = [game.Game]
+    TYPES = [games.template.RL]
     BASE_CONFIG = {}
 
 
@@ -42,10 +42,10 @@ class TestRL(unittest.TestCase, ModuleTest):
     """
 
     TYPES = [
-        RL.Logic,
-        RL.CartPole,
-        gym_wrapper(cartpole.CartPoleEnv, base=RL.template.RL),
-        gym_wrapper(mountain_car.MountainCarEnv, base=RL.template.RL),
+        games.Logic,
+        games.CartPole,
+        games.gym_wrapper(cartpole.CartPoleEnv),
+        games.gym_wrapper(mountain_car.MountainCarEnv),
     ]
     BASE_CONFIG = {}
